Renamed 'dashboard' to 'cashflow' to be more accurate added single categorization to generic transaction view used, which is primarily used in the cashflow module genericized the cashflow modules to make their code more common with each other, making them more extensible adding transaction deletion tools Creating early category model

+++ src/output/TransactionList.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import TransactionModel from '../models/Transactions.jsx';
+import CategoryModel from '../models/Categories.jsx';
 
 export default class TransactionList extends React.Component{
 	render(){
@@ -12,6 +13,7 @@ export default class TransactionList extends React.Component{
 							<th>Description</th>
 							<th>Debit</th>
 							<th>Credit</th>
+							<th>Category</th>
 							<th></th>
 						</tr>
 					</thead>
@@ -29,7 +31,7 @@ export class TransactionRows extends React.Component{
 	}
 	render(){
 		console.log("Rendering Transaction Rows",typeof this.props.transactions,this.props.transactions);
-		if(typeof this.props.transactions !== "array" && typeof this.props.transactions !== "object"){
+		if(typeof this.props.transactions !== "object"){
 			return (
 				<tr>
 					<td colSpan="4">
@@ -40,27 +42,52 @@ export class TransactionRows extends React.Component{
 		}
 		console.log(this.deleted)
 		return this.props.transactions
-			.filter(record => this.deleted.indexOf(record.id) == -1)
+			.filter(record => this.deleted.indexOf(record.id) === -1)
 			.map(this.renderRow.bind(this));
 	}
 
-	renderRow(row,key){
+	renderRow(row){
 		let moneyStyle = {
 			textAlign:'right',
 			fontFamily: 'Ubunto Mono,monospace,mono'
 		};
 		return (
-			<tr key={key}>
+			<tr key={row.id}>
 				<td>{row.date}</td>
 				<td>{row.description}</td>
 				<td style={moneyStyle}>{row.debit.toFixed(2)}</td>
 				<td style={moneyStyle}>{row.credit.toFixed(2)}</td>
+				<td>
+					<select onChange={this.categorize.bind(this,row.id)}>
+						{this.getCategoryOptions(row.categories)}
+					</select>
+				</td>
 				<td><button onClick={this.deleteTransaction.bind(this,row)}>Delete</button></td>
 			</tr>
 		);
 	}
 
-	deleteTransaction(row){
+	categorize(transaction,event){
+		return CategoryModel.categorizeTransaction(event.target.value,transaction)
+			.then(TransactionModel.fetchTransactions);
+	}
+
+	getCategoryOptions(categories){
+		if(!categories) categories = [];
+		const opts = CategoryModel.getCategories().sort((cat1,cat2) => ('' + cat1.name).localeCompare(cat2.name)).map(cat => {
+			const props = {value:cat.id};
+			if( categories.map(record => record.id).indexOf(cat.id) !== -1)
+				props.selected = "selected";
+			return (<option {...props}>{cat.name}</option>);
+		});
+		opts.unshift((<option value={null}>- Uncategorized -</option>));
+		return opts;
+	}
+
+	deleteTransaction(row,event){
+		if(event.target.innerHTML === "Delete"){
+			return event.target.innerHTML = "Confirm Delete?";
+		}
 		TransactionModel.deleteTransaction(row.id);
 		this.deleted.push(row.id);
 	}
